requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
autopep8==2.0.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-cors-headers==3.13.0
django-image-optimizer==1.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
exceptiongroup==1.0.0
Faker==15.1.1
h11==0.14.0
idna==3.4
joblib==1.2.0
numpy==1.23.4
optimize-images==1.5.1
outcome==1.2.0
packaging==21.3
pandas==1.5.1
piexif==1.1.3
Pillow==9.2.0
pycodestyle==2.9.1
PyJWT==2.6.0
pymongo==4.3.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-resize-image==1.1.20
pytz==2022.5
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
selenium==4.5.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
threadpoolctl==3.1.0
tinify==1.6.0
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
ujson==5.5.0
urllib3==1.26.12
watchdog==2.1.9
webdriver-manager==3.8.4
whitenoise==6.2.0
wsproto==1.2.0
